services/user-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy>=2.0
alembic
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg[binary]==3.2.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
email-validator==2.2.0
pytest
httpx
pytest-cov
